50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,17 +5,17 @@
 */
 int main(void)
 {
-	unsigned long long int x1 = 0;
-	unsigned long long int x2 = 1;
+	unsigned long int x1 = 0;
+	unsigned long int x2 = 1;
 	int i;
-	unsigned long long int sum = 0;
+	unsigned long  int sum = 0;
 
 	for (i = 2; i < 59; i++)
 	{
 		sum = x1 + x2;
 		x1 = x2;
 		x2 = sum;
-		printf("%llu ", sum);
+		printf("%lu ", sum);
 		if (i  < 58)
 			printf(", ");
 	}
